requirements.txt
boto3>=1.35.0
streamlit>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.9.0
streamlit-mic-recorder>=0.0.8
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pyjwt>=2.9.0
